w3c-reports/12092014/mac-wk2/runned-tests/FileAPI/Blob-constructor-expected.txt

CONSOLE MESSAGE: line 65: ReferenceError: Can't find variable: test_blob
Discussion is ongoing that will affect a number of the following tests.


Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: test_blob
